Blob storage: align blob client code (#1215)
