+++ extensions/azure/blobstorage/blob-core/src/main/java/org/eclipse/dataspaceconnector/azure/blob/core/BlobStoreCoreExtension.java
@@ -26,7 +26,7 @@ import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 public class BlobStoreCoreExtension implements ServiceExtension {
 
     @EdcSetting
-    public static final String EDC_BLOBSTORE_ENDPOINT = "edc.blobstore.endpoint";
+    public static final String EDC_BLOBSTORE_ENDPOINT_TEMPLATE = "edc.blobstore.endpoint.template";
 
     @Override
     public String name() {
@@ -35,9 +35,10 @@ public class BlobStoreCoreExtension implements ServiceExtension {
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        var blobstoreEndpoint = context.getSetting(EDC_BLOBSTORE_ENDPOINT, null);
+        var blobstoreEndpointTemplate = context
+                .getSetting(EDC_BLOBSTORE_ENDPOINT_TEMPLATE, "https://%s.blob.core.windows.net");
 
-        var blobStoreApi = new BlobStoreApiImpl(context.getService(Vault.class), blobstoreEndpoint);
+        var blobStoreApi = new BlobStoreApiImpl(context.getService(Vault.class), blobstoreEndpointTemplate);
         context.registerService(BlobStoreApi.class, blobStoreApi);
     }
 }
