+++ extensions/azure/blobstorage/blob-core/src/main/java/org/eclipse/dataspaceconnector/azure/blob/core/api/BlobStoreApi.java
@@ -15,6 +15,7 @@
 package org.eclipse.dataspaceconnector.azure.blob.core.api;
 
 import com.azure.storage.blob.models.BlobItem;
+import org.eclipse.dataspaceconnector.azure.blob.core.adapter.BlobAdapter;
 
 import java.time.OffsetDateTime;
 import java.util.List;
@@ -36,4 +37,6 @@ public interface BlobStoreApi {
     String createAccountSas(String accountName, String containerName, String racwxdl, OffsetDateTime expiry);
 
     byte[] getBlob(String account, String container, String blobName);
+
+    BlobAdapter getBlobAdapter(String accountName, String containerName, String blobName, String sharedKey);
 }
