+++ extensions/azure/blobstorage/blob-core/src/test/java/org/eclipse/dataspaceconnector/azure/blob/core/api/BlobStoreApiImplTest.java
@@ -1,0 +1,37 @@
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - Initial implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.azure.blob.core.api;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatNoException;
+import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createAccountName;
+import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createBlobName;
+import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createContainerName;
+import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createSharedKey;
+
+class BlobStoreApiImplTest {
+
+    @Test
+    void getBlobAdapter_succeeds() {
+        var service = new BlobStoreApiImpl(null, "https://%s.blob.core.windows.net");
+        assertThatNoException()
+                .isThrownBy(() -> service.getBlobAdapter(
+                        createAccountName(),
+                        createContainerName(),
+                        createBlobName(),
+                        createSharedKey()));
+    }
+}
