--- extensions/azure/data-plane/common/build.gradle.kts
@@ -1,31 +1,0 @@
-/*
- *  Copyright (c) 2022 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    api(project(":extensions:data-plane:data-plane-spi"))
-    implementation(project(":common:util"))
-}
-
-publishing {
-    publications {
-        create<MavenPublication>("data-plane-azure-common") {
-            artifactId = "data-plane-azure-common"
-            from(components["java"])
-        }
-    }
-}
