+++ extensions/azure/data-plane/storage/build.gradle.kts
@@ -12,32 +12,24 @@
  *
  */
 
-val storageBlobVersion: String by project;
+val storageBlobVersion: String by project
 val jodahFailsafeVersion: String by project
-val jupiterVersion: String by project
-val faker: String by project
 
 plugins {
     `java-library`
-    `java-test-fixtures`
 }
 
 
 dependencies {
-    implementation(project(":extensions:azure:data-plane:common"))
+    api(project(":extensions:data-plane:data-plane-spi"))
+    api(project(":extensions:azure:blobstorage:blob-core"))
     implementation(project(":common:util"))
     implementation("com.azure:azure-storage-blob:${storageBlobVersion}")
     implementation("net.jodah:failsafe:${jodahFailsafeVersion}")
 
     testImplementation(testFixtures(project(":extensions:azure:azure-test")))
+    testImplementation(testFixtures(project(":extensions:azure:blobstorage:blob-core")))
     testImplementation(testFixtures(project(":common:util")))
-
-    testFixturesApi(project(":extensions:data-plane:data-plane-spi"))
-    testFixturesImplementation(project(":common:util"))
-
-    testFixturesImplementation("org.junit.jupiter:junit-jupiter-api:${jupiterVersion}")
-    testFixturesRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:${jupiterVersion}")
-    testFixturesImplementation("com.github.javafaker:javafaker:${faker}")
 }
 
 publishing {
