+++ extensions/azure/data-plane/storage/src/test/java/org/eclipse/dataspaceconnector/azure/dataplane/azurestorage/pipeline/AzureDataSourceToDataSinkTest.java
@@ -16,18 +16,23 @@ package org.eclipse.dataspaceconnector.azure.dataplane.azurestorage.pipeline;
 
 import com.github.javafaker.Faker;
 import net.jodah.failsafe.RetryPolicy;
-import org.eclipse.dataspaceconnector.azure.dataplane.azurestorage.adapter.BlobAdapter;
-import org.eclipse.dataspaceconnector.azure.dataplane.azurestorage.adapter.BlobAdapterFactory;
-import org.eclipse.dataspaceconnector.azure.dataplane.azurestorage.pipeline.AzureStorageTestFixtures.FakeBlobAdapter;
+import org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures;
+import org.eclipse.dataspaceconnector.azure.blob.core.adapter.BlobAdapter;
+import org.eclipse.dataspaceconnector.azure.blob.core.api.BlobStoreApi;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.junit.jupiter.api.Test;
 
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.UUID;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
 import static java.lang.String.format;
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -53,7 +58,7 @@ class AzureDataSourceToDataSinkTest {
      */
     @Test
     void transfer_success() {
-        var fakeSourceFactory = mock(BlobAdapterFactory.class);
+        var fakeSourceFactory = mock(BlobStoreApi.class);
         when(fakeSourceFactory.getBlobAdapter(
                 sourceAccountName,
                 sourceContainerName,
@@ -68,11 +73,11 @@ class AzureDataSourceToDataSinkTest {
                 .blobName(fakeSource.name)
                 .requestId(requestId)
                 .retryPolicy(new RetryPolicy<>())
-                .blobAdapterFactory(fakeSourceFactory)
+                .blobStoreApi(fakeSourceFactory)
                 .monitor(monitor)
                 .build();
 
-        var fakeSinkFactory = mock(BlobAdapterFactory.class);
+        var fakeSinkFactory = mock(BlobStoreApi.class);
         when(fakeSinkFactory.getBlobAdapter(
                 sinkAccountName,
                 sinkContainerName,
@@ -85,7 +90,7 @@ class AzureDataSourceToDataSinkTest {
                 .containerName(sinkContainerName)
                 .sharedKey(sinkSharedKey)
                 .requestId(requestId)
-                .blobAdapterFactory(fakeSinkFactory)
+                .blobStoreApi(fakeSinkFactory)
                 .executorService(executor)
                 .monitor(monitor)
                 .build();
@@ -106,7 +111,7 @@ class AzureDataSourceToDataSinkTest {
         var blobAdapter = mock(BlobAdapter.class);
         when(blobAdapter.getBlobName()).thenReturn(fakeSource.name);
         when(blobAdapter.openInputStream()).thenThrow(new RuntimeException(errorMessage));
-        var fakeSourceFactory = mock(BlobAdapterFactory.class);
+        var fakeSourceFactory = mock(BlobStoreApi.class);
         when(fakeSourceFactory.getBlobAdapter(
                 sourceAccountName,
                 sourceContainerName,
@@ -121,11 +126,11 @@ class AzureDataSourceToDataSinkTest {
                 .blobName(fakeSource.name)
                 .requestId(requestId)
                 .retryPolicy(new RetryPolicy<>())
-                .blobAdapterFactory(fakeSourceFactory)
+                .blobStoreApi(fakeSourceFactory)
                 .monitor(monitor)
                 .build();
 
-        var fakeSinkFactory = mock(BlobAdapterFactory.class);
+        var fakeSinkFactory = mock(BlobStoreApi.class);
         when(fakeSinkFactory.getBlobAdapter(
                 sinkAccountName,
                 sinkContainerName,
@@ -138,7 +143,7 @@ class AzureDataSourceToDataSinkTest {
                 .containerName(sinkContainerName)
                 .sharedKey(sinkSharedKey)
                 .requestId(requestId)
-                .blobAdapterFactory(fakeSinkFactory)
+                .blobStoreApi(fakeSinkFactory)
                 .executorService(executor)
                 .monitor(monitor)
                 .build();
@@ -155,7 +160,7 @@ class AzureDataSourceToDataSinkTest {
     void transfer_whenSinkFails_fails() throws Exception {
 
         // source completes normally
-        var fakeSourceFactory = mock(BlobAdapterFactory.class);
+        var fakeSourceFactory = mock(BlobStoreApi.class);
         when(fakeSourceFactory.getBlobAdapter(
                 sourceAccountName,
                 sourceContainerName,
@@ -170,14 +175,14 @@ class AzureDataSourceToDataSinkTest {
                 .blobName(fakeSource.name)
                 .requestId(requestId)
                 .retryPolicy(new RetryPolicy<>())
-                .blobAdapterFactory(fakeSourceFactory)
+                .blobStoreApi(fakeSourceFactory)
                 .monitor(monitor)
                 .build();
 
         // sink endpoint raises an exception
         var blobAdapter = mock(BlobAdapter.class);
         when(blobAdapter.getOutputStream()).thenThrow(new RuntimeException());
-        var fakeSinkFactory = mock(BlobAdapterFactory.class);
+        var fakeSinkFactory = mock(BlobStoreApi.class);
         when(fakeSinkFactory.getBlobAdapter(anyString(), anyString(), anyString(), anyString()))
                 .thenReturn(blobAdapter);
 
@@ -186,7 +191,7 @@ class AzureDataSourceToDataSinkTest {
                 .containerName(sinkContainerName)
                 .sharedKey(sinkSharedKey)
                 .requestId(requestId)
-                .blobAdapterFactory(fakeSinkFactory)
+                .blobStoreApi(fakeSinkFactory)
                 .executorService(executor)
                 .monitor(monitor)
                 .build();
@@ -194,4 +199,30 @@ class AzureDataSourceToDataSinkTest {
         assertThat(dataSink.transfer(dataSource).get().failed()).isTrue();
     }
 
+    static class FakeBlobAdapter implements BlobAdapter {
+        final String name = faker.lorem().characters();
+        final String content = faker.lorem().sentence();
+        final long length = faker.random().nextLong(1_000_000_000_000_000L);
+        final ByteArrayOutputStream out = new ByteArrayOutputStream();
+
+        @Override
+        public OutputStream getOutputStream() {
+            return out;
+        }
+
+        @Override
+        public InputStream openInputStream() {
+            return new ByteArrayInputStream(content.getBytes(UTF_8));
+        }
+
+        @Override
+        public String getBlobName() {
+            return name;
+        }
+
+        @Override
+        public long getBlobSize() {
+            return length;
+        }
+    }
 }
