+++ settings.gradle.kts
@@ -68,8 +68,6 @@ include(":extensions:azure:blobstorage")
 include(":extensions:azure:blobstorage:blob-core")
 include(":extensions:azure:blobstorage:blob-provision")
 include(":extensions:azure:blobstorage:blob-data-operator")
-include(":extensions:azure:data-plane:common")
-include(":extensions:azure:data-plane:storage")
 include(":extensions:azure:events")
 include(":extensions:azure:events-config")
 include(":extensions:azure:azure-test")
@@ -136,7 +134,6 @@ include(":extensions:data-plane-selector:selector-api")
 include(":extensions:data-plane-selector:selector-core")
 include(":extensions:data-plane-selector:selector-store")
 include(":extensions:data-plane-selector:selector-client")
-include(":extensions:azure:data-plane:common")
 include(":extensions:azure:data-plane:storage")
 include(":extensions:data-plane:data-plane-api")
 include(":extensions:data-plane:integration-tests")
